Move nud/led/lbp methods to Parser class
It'll be easier to do the Tokenizer if Token was just a small class holding data that the Tokenizer just extracted. Then a simple enum switch can be done all contained inside the Parser code to get it done. These 3 methods (nud, led, lbp) are protected so then subclasses can override or extend this behavior.